+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Last Changes
 
+- [#2](https://github.com/LaxarJS/grunt-init-laxar-widget/issues/2): Updated `integration` field in `widget.json` template
 - [#1](https://github.com/LaxarJS/grunt-init-laxar-widget/issues/1): Removed `version` field from `widget.json` template